requirements.txt
# LangChain ecosystem
langchain==0.1.0
langchain-core
langchain-community==0.0.13
langchain-groq==0.0.1
langchain_openai
langgraph

# Document loaders & processing
pypdf==3.17.4
pymupdf
python-docx==1.1.0
openpyxl==3.1.2
unstructured[local-inference]==0.11.8

# Embeddings & vector stores
sentence-transformers==2.3.1
faiss-cpu
chromadb
typesense
torch>=2.0.0
numpy>=1.24.0

# Core web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6

# Database & env
psycopg2-binary==2.9.9
python-dotenv==1.0.0

# Utilities
pathlib2==2.3.7.post1